backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.10",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.20",
    "langchain-community>=0.3.29",
    "langchain-deepseek>=0.1.4",
    "langchain-google-genai>=2.1.10",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.33",
    "langchain-perplexity>=0.1.2",
    "numpy>=2.3.3",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-pip>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "unstructured[all-docs]>=0.18.15",
    "streamlit>=1.49.1",
    "uvicorn>=0.36.0",
    "drf-api-logger>=1.1.21",
    "langchain-ollama>=0.3.10",
]
